gateway-server/requirements.txt
fastapi==0.78.0
fuzzywuzzy==0.18.0
jsonschema==3.2.0
pymodbus==2.5.3
pyModbusTCP==0.2.0
uvicorn==0.18.2
PyYAML==6.0